+++ rtengine/matrix33.h
@@ -16,7 +16,7 @@ namespace rtengine {
 // color (r,g,b) triplets between different color spaces
 class Matrix33 {
 
-		double data[3][3];
+		float data[3][3];
 
 	public:
 		// initializes the matrix with the array passed
@@ -25,18 +25,18 @@ class Matrix33 {
 
 		// applies transformation on the given (r,g,b) (column) vector
 		// result is written back to the variables passed
-        void transform (double& r, double& g, double& b);
+        void transform (float& r, float& g, float& b);
         void transform (unsigned short& r, unsigned short& g, unsigned short& b);
 
 		// the same, result is stored separately
-        void transform (double r, double g, double b, double& nr, double& ng, double& nb);
+        void transform (float r, float g, float b, float& nr, float& ng, float& nb);
         void transform (unsigned short r, unsigned short g, unsigned short b, unsigned short& nr, unsigned short& ng, unsigned short& nb);
 
 		// returns inverse of the transformation matrix
 		Matrix33 inverse ();
 
 		// multiplies this matrix from the given one from the right
-		void multiply (const double (*m)[3]);
+		void multiply (const float (*m)[3]);
 };
 
 }
